refactor: create project status tag component
