+++ src/components/Tooltip.tsx
@@ -1,22 +1,6 @@
 import { useMapReady } from '@ugrc/utilities/hooks';
 import { useEffect, useRef, useState } from 'react';
-import { tv } from 'tailwind-variants';
-import { Status } from './ProjectStatus';
-
-const tagStyles = tv({
-  base: 'flex max-w-fit cursor-default items-center gap-1 rounded-md border px-3 py-0.5 text-xs text-white',
-  variants: {
-    status: {
-      draft: 'border-gray-200 bg-zinc-500',
-      proposed: 'border-gray-200 bg-zinc-800',
-      current: 'border-gray-200 bg-sky-600',
-      'pending completed': 'border-gray-200 bg-yellow-500',
-      completed: 'border-gray-200 bg-green-700',
-      cancelled: 'border-gray-200 bg-red-700',
-      undefined: 'border-gray-200 bg-gray-200',
-    },
-  },
-});
+import { ProjectStatusTag } from './ProjectStatusTag';
 
 type GraphicAttributes = {
   Project_ID: number;
@@ -77,14 +61,12 @@ export const Tooltip = ({ view, layers }: { view: __esri.MapView; layers: __esri
     });
   }, [isReady, layers, view]);
 
-  const status = (hoverProject?.Status ?? hoverProject?.StatusDescription)?.toLocaleLowerCase() as Status;
-
   return (
     <div ref={tooltipNode} className="absolute z-10 h-20 w-56 rounded border bg-white opacity-0 transition-opacity">
       <div className="p-2 text-sm">
         <h3 className="truncate text-sm font-bold">{hoverProject?.Title}</h3>
         <p>Project: {hoverProject?.Project_ID}</p>
-        <div className={tagStyles({ status: status })}>{hoverProject?.Status ?? hoverProject?.StatusDescription}</div>
+        <ProjectStatusTag status={hoverProject?.Status ?? hoverProject?.StatusDescription ?? ''} />
       </div>
     </div>
   );
