+++ src/components/index.ts
@@ -8,6 +8,7 @@ export * from './MapContainer.tsx';
 export * from './ProjectGeneral.tsx';
 export * from './ProjectSpecific.tsx';
 export * from './ProjectStatus.tsx';
+export * from './ProjectStatusTag.tsx';
 export * from './ReferenceData.tsx';
 export * from './Tooltip.tsx';
 export * from './utils.ts';
